Committing stuff now cos it works and I don't want it to break
I have set up scripts to update the database for a new day and to repopulate the entire thing if needed. I have set a user agent on my local machine to run the daily script once per day (shocker) at 12:01am so that I can also schedule the model to make predictions with new data when this has completed.